[test_youtube_misc] Move YoutubeIE.extract_id test into separate module
